Implement pan-zoom and content fitting without object-fit
When using the object-fit CSS property, panning and zooming needs to be
implemented via the object-position, left, top, width and height CSS
properties. Animating these does not result in a smooth movement.

This patch implements the fitting via setting the content's width
and height to auto together with setting min/max-width/-height to 100%,
depending on the fitting type. This allows to utilize the smoothly
animated CSS transforms for panning and zooming.