Migrate src/.eslintrc.yaml to src/.eslintrc.js to add __dirname support
The @typescript-eslint/parser option tsconfigRootDir is interpreted
relative to the first .eslintrc.yaml file in the chain, not relative to
the .eslintrc.yaml file it is defined in. This causes all sorts of
issue when multiple .eslintrc.yaml files are combined. Switchting the
root file to .eslintrc.js allows to use __dirname which will be
evaluated in the file it is used and therefore results in the correct
paths everywhere.