Make the fit option actually work
The fit options did not work for most cases, especially when the content
was small compared to the container or when the aspect ratio of the
container changed past the aspect ratio of the content.

The new system tracks size changes of the container using the
(polyfilled) ResizeObserver system and adjusts certain CSS properties
accordingly. The approach is still reasonably lightweight. A layout
change may only be triggered, when the aspect ratios of the container
and the content cross during size adjustments.