Fix display issues arising from fractional scaling
The previous implementation combined absolute positioning and CSS
transforms to center the content in the slide. However, due to
CSS rounding errors, the positioning of the content was sometimes
off by one pixel. A single column or row of pixels of the background
became visible in this case, which looked quite ugly if the content and
the background had very different colors.

This patch switches the centering over to CSS Flexbox layout which seems
a lot more stable in this regard. Some necessary modifications to the
`pan-zoom` animation have applied as well.