Fix dynamic creation of `<hilbert-gallery-viewer>` element
The previous constructor of HilbertGalleryViewer inserted an
additional wrapper into the container element supplied as argument.
This triggered an exception in the call
`document.createElement('hilbert-gallery-viewer')` because newly created
elements need to be empty.

This patch adds an additional `wrap` parameter to the constructor of
HilbertGalleryViewer to signal if this addtional layer should be
inserted. The flag is off for calls from the HilbertGalleryViewerElement
constructor.